plus up
Former-commit-id: 54d5f44a65b95cab348f252e2093857bdf22c14b

+++ Experimental Area/ft_printf/inc/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: adebray <adebray@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2013/12/16 20:25:00 by adebray           #+#    #+#             */
-/*   Updated: 2013/12/21 18:28:47 by adebray          ###   ########.fr       */
+/*   Updated: 2013/12/22 08:17:38 by adebray          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ char				*ft_cutstring(char *from, char *to);
 int					ft_isdigit(int c);
 int					ft_atoi(const char *str);
 int					ft_putoctal(unsigned int decimal);
-int					ft_puthexa(unsigned int decimal);
+int					ft_puthexa(unsigned long int decimal);
 int					ft_put_sheum_hexa(unsigned int decimal);
